serve: isolate shutdown tests; health-check controller on shutdown
- Add test_shutdown.py with shutdown-related tests moved from test_standalone.
- Keep driver-reconnect regression tests commented for baseline CI.
- Pass _health_check_controller=True in shutdown() and shutdown_async().
- Register test_shutdown in serve tests BUILD (enormous).

Made-with: Cursor
